Move trapi and ARSClient modules to lib/
