+++ TranslatorService.mjs
@@ -2,7 +2,7 @@
 
 import * as cmn from './common.mjs';
 import * as arsmsg from './ARSMessages.mjs';
-import * as trapi from './trapi.mjs';
+import * as trapi from './lib/ARSClient.mjs';
 
 export { TranslatorService };
