+++ lib/trapi.mjs
