+++ mock/main.mjs
@@ -2,7 +2,7 @@
 
 import { SERVER_CONFIG } from '../config.mjs';
 import * as ars from './ars.mjs';
-import * as trapi from '../trapi.mjs';
+import * as trapi from '../lib/trapi.mjs';
 import * as server from '../server.mjs';
 
 server.start(ars, trapi, SERVER_CONFIG);
