+++ utilities/node/retainPKs.mjs
@@ -1,6 +1,6 @@
 import * as fs from 'node:fs'
 import * as cmn from '../../common.mjs';
-import { ARSClient } from '../../ARSClient.mjs';
+import { ARSClient } from '../../lib/ARSClient.mjs';
 
 const configRoot = '../../configurations';
 const env = process.argv[2];
